+++ NEWS.md
@@ -1,3 +1,8 @@
+# watcher 0.1.2
+
+* Adds `$get_path()` and `$is_running()` methods to the `Watcher` R6 class.
+  + Use these rather than the fields `path` and `running`, as they have been made private.
+
 # watcher 0.1.1
 
 * Updates bundled 'libfswatch' source package to 1.19.0-dev.